Have made a working class and new script for the pic_file changer.
Now I am trying to make to usable with both pictures and videos.
At this point I need to fix when there is no match found. Need to only
print out one 'warning'. As well should not continue the script and
re-ask for a new date from the user.

I also was thinking of displaying the directory content when the script
is loaded to give a nice visual.